Separate actionTypes into independent module
